requirements.txt
bayesian-torch
matplotlib
pandas
seaborn
jupyter
ipykernel
hydra-core[all]
